# AI客服/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ruitalk Seller Terminal API",
    "version": "1.0.0",
    "description": "AI-powered multilingual customer service system API",
    "contact": {
      "name": "Ruitalk Team",
      "email": "team@ruitalk.com"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local Development"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Alternate"
    }
  ],
  "paths": {
...
    "/api/v1/shop/shops": {
      "post": {
        "summary": "创建店铺",
        "description": "创建店铺",
        "operationId": "post__api_v1_shop_shops",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "summary": "获取店铺列表",
        "description": "获取店铺列表",
        "operationId": "get__api_v1_shop_shops",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/shop/shops/{shop_id}": {
      "get": {
        "summary": "获取单个店铺",
        "description": "获取单个店铺",
        "operationId": "get__api_v1_shop_shops_shop_id",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: shop_id"
          }
        ]
      },
      "put": {
        "summary": "更新店铺",
        "description": "更新店铺",
        "operationId": "put__api_v1_shop_shops_shop_id",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: shop_id"
          }
        ]
      },
      "delete": {
        "summary": "删除店铺",
        "description": "删除店铺",
        "operationId": "delete__api_v1_shop_shops_shop_id",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: shop_id"
          }
        ]
      }
    },
    "/api/v1/shop/shops/{shop_id}/test": {
      "post": {
        "summary": "测试店铺连接",
        "description": "测试店铺连接",
        "operationId": "post__api_v1_shop_shops_shop_id_test",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: shop_id"
          }
        ]
      }
    },
    "/api/v1/shop/platforms": {
      "get": {
        "summary": "获取支持的平台列表",
        "description": "获取支持的平台列表",
        "operationId": "get__api_v1_shop_platforms",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/shop/products": {
      "post": {
        "summary": "创建商品",
        "description": "创建商品",
        "operationId": "post__api_v1_shop_products",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "summary": "获取商品列表",
        "description": "获取商品列表",
        "operationId": "get__api_v1_shop_products",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/shop/products/{product_id}": {
      "get": {
        "summary": "获取单个商品（含SKU）",
        "description": "获取单个商品（含SKU）",
        "operationId": "get__api_v1_shop_products_product_id",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: product_id"
          }
        ]
      },
      "put": {
        "summary": "更新商品",
        "description": "更新商品",
        "operationId": "put__api_v1_shop_products_product_id",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: product_id"
          }
        ]
      },
      "delete": {
        "summary": "删除商品",
        "description": "删除商品",
        "operationId": "delete__api_v1_shop_products_product_id",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: product_id"
          }
        ]
      }
    },
...
    "/api/v1/shop/skus/{sku_id}": {
      "put": {
        "summary": "更新SKU",
        "description": "更新SKU",
        "operationId": "put__api_v1_shop_skus_sku_id",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: sku_id"
          }
        ]
      },
      "delete": {
        "summary": "删除SKU",
        "description": "删除SKU",
        "operationId": "delete__api_v1_shop_skus_sku_id",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: sku_id"
          }
        ]
      }
    },
    "/api/v1/shop/collect": {
      "post": {
        "summary": "采集商品",
        "description": "采集商品",
        "operationId": "post__api_v1_shop_collect",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/shop/publish": {
      "post": {
        "summary": "批量刊登商品",
        "description": "批量刊登商品",
        "operationId": "post__api_v1_shop_publish",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/shop/shop-products": {
      "get": {
        "summary": "获取店铺商品列表",
        "description": "获取店铺商品列表",
        "operationId": "get__api_v1_shop_shop-products",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/v1/shop/pricing-rules/{rule_id}": {
      "delete": {
        "summary": "删除定价规则",
        "description": "删除定价规则",
        "operationId": "delete__api_v1_shop_pricing-rules_rule_id",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: rule_id"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token"
      }
    },
...
  }
}